chore: Add StaticStates wrapper (#377)
* Update story sorting to allow multiple-tier sorting vs global sorting

This creates a wrapper for stories to allow for treating pseudo-selectors like `:hover` or `:focus` as static states that are controllable as classes rather than the default pseudo-selectors.

```tsx
.add('States', () => {
  <StaticStates>
    <Button className="focus">Focus</Button>
  </StaticStates>
});
```

This code will render a button with focus styling always applied. This is useful for enumerating states for design or validating states in visual regression tests.

To enable this functionality, Components must use the Canvas-kit `styled` function instead of the one from Emotion.