feat(tooltip): Refactor to a simpler API (#528)
* Add unit tests
* Add Cypress specifications
* Add visual tests
* Upgrade Popper to v2 and use canvas-kit Popper component instead of materia-ui's

This change makes it easier to create accessible tooltips. It is now possible to do this:

```tsx
<Tooltip title="Close">
  <IconButton variant={IconButton.Variant.Circle} icon={xIcon} aria-label="Close" />
</Tooltip>
```

BREAKING CHANGE: Tooltip API was updated to make it easier to consume.
The original `Tooltip` did little more than add a `role="tooltip"` to a styled component. The original tooltip is now exported as `TooltipContainer` to make it easier to migrate without rewriting all tooltips. Imports will have to be updated to use the old API:
**Before:**
```ts
import { Tooltip } from '@workday/canvas-kit-react-tooltip'
```
**After:**
```ts
import { TooltipContainer as Tooltip } from '@workday/canvas-kit-react-tooltip'
```
Also with this change, the tooltip no longer gets the role `tooltip` and must be added manually.

BREAKING CHANGE: `Popper` was changed to a Functional Component with a forwarded ref. If you passed a `ref` object to `Popper` before, it will now point to the element rather than the `Popper` instance.