fix(modal): Add exception to modal enter animation (#867)
* fix(modal): Add exception to modal enter animation

This fixes an issue where the Popup adapter is being used. It will flash because the adapter will remove the containing element for popups and show again. The animation makes detach/attach cycle a visible flashing. This is already added to the Popup.tsx file, we just missed the Modal.