fix(tooltip): Add delay to Tooltip show and hide (#1339)
This PR introduces the default delay time of 300ms for showing a tooltip, and adds the ability to control the delay time for showing and hiding the tooltip.

[category:Components]

Release Note:
This change could cause visual regression tests to fail if a screen shot is taken expecting a tooltip to show immediately. Your visual regression will either have to add an explicit wait of 300ms, or change the delay to 1ms only under test.