fix: Revert uuid upgrade to get releases working [skip release] (#1371)
Fix broken releases introduced by #1367 that unintentionally removed `@types/node` from Typescript which is causing `process.env` to be undeclared. Also updating the `yarn typecheck` scripts to catch this type of error. The PR job does not run `yarn build`, but runs `yarn typecheck`. Before the whole project was typechecked as a single project which worked because the `css` module included storyfiles using `@storybook/react` which included `webpack` which included definitions files for `process` and thus this PR was able to be merged even though a `yarn build` would fail. `yarn typecheck` is now being run the same way as `yarn build` is. This should prevent this type of issue.

[category:Revert]