feat!: Increase max width of Status Indicator, add prop and examples (#1329)
Fixes: #1317 

Status Indicators currently truncate when they reach their max width of `150px`. After receiving
requests to increase this, we have upped it by 25% to `200px`.

I've also added a prop (`maxWidth`) to configure this if you need to. I've added a new story to show an example of this, and with an `OverflowTooltip`

[category:Components]

### BREAKING CHANGES
Status Indicators currently truncate when they reach their max width of `150px`. After receiving
requests to increase this, we have upped it by 25% to `200px`.