chore!: Upgrade to React 17 and refactor our css prop (#1409)
- Upgrade React and React Dom to 17
- Remove use `css` prop and `/** @jsx jsx */` pragma in examples and internal components
- Refactor components to use styled components where applicable

Fixes: https://github.com/Workday/canvas-kit/issues/1334

[category:Dependencies]

### BREAKING CHANGES
Canvas Kit will require teams to be on version ^17.x or at least 16.14 for backwards compatibility and we will remove the jsx pragma and use of css props. This does not mean consumers can’t use the css prop, just that internally we will stop using it. This would require teams to upgrade other parts of their app  including  React, potentially Babel and any other side effects that come with it.