feat(modal): Add support for scrolling a modal (#1259)
Fixes #790

This PR changes the structure of Modals slightly by removing the extra `div` in the `Modal.Body` and `Popup.Body` and instead adds the type that `Card.Body` adds manually. This allows the `Modal.Body` to be more easily styled, but may be a breaking change. This change also allows overflow to be applied to Modals.

[category:Components]