feat(preview): Add Pill component (#1542)
Fixes: (https://github.com/Workday/canvas-kit/issues/538)

[category:Components]

Release Note:
We've added a new Pill component. Pills are used to visually label objects on a page for quick recognition. They’re offered as both static, read-only and interactive elements. They allow users to input a section, filter a list or table, or label information to help with scanning and organization.