test(select): Reduce sensitivity of scrolling test [skip release] (#1613)
Reduces the sensitivity of a couple Cypress tests for the Preview Select which were looking at `window.scrollY`. These tests were failing locally (`expecting 2120 to equal 2120.5`) but not on the CI, which may be why we didn't notice this until now. We suspect Chrome was updated recently to support a fractional `scrollY`.

In any event, a half-pixel of scrolling doesn't even seem to produce visible scrolling. I've relaxed the tests to chop off the fractional pixel when making the comparison.

[category:Components]