fix: Remove spellcheck from togglable password field (#1782)
Remove spellcheck on the password input formik example. Without this attribute the password is sent to the browser spellcheck service in plaintext as soon as the type attribute is toggled. 

This is pretty easy to verify without inspecting the network traffic. Without this change you will see the red squiggly line right after clicking the eye icon, after this change there is no squiggly.

[category:Documentation]