docs: Make release category easier to change in PR template (#1804)
We've been using the release category badge for a while, and it is difficult to parse and change the category. It is also harder to read the category in the badge due to lower contrast. This change updates the pull request template to make "Release Category" be a regular section.

[category:Documentation]