fix(popup): Fix return focus timing for tooltips and focus events (#1806)
Fixes: #1805

There is a bug where if a popup is closed and `useReturnFocus` is used for the popup and the popup target has a `Tooltip`, the tooltip is not shown when the popup is closed. This is due to the synchronous nature of `useReturnFocus`. `useReturnFocus` uses a synchronous focus change to handle `useFocusRedirect` correctly. Additional event state tracking was added to detect when we want the browser to change focus on the popup target or when we want the browser to trigger a focus event on the popup target.

[category:Components]