ci: Close related issues with pull requests [skip release] (#1817)
Close related issues when the pull request is closed regardless of the target branch of the pull request. Github already closes issues with pull requests targeting the default branch, but Canvas Kit supports multiple support branches. An issue is considered closed when it is merged into `support`, for example. This change decreases the overhead of maintaining Canvas Kit Github.

[category:Infrastruture]