fix!: Popup close buttons no longer interfere with forms (#1861)
Update the default `type` of popup `CloseIcon` and `CloseButton` components to `button` instead of `submit`. Also, we added a form modal example.

[category:Components]

### BREAKING CHANGES
We've updated all popup `CloseIcon` and `CloseButton` to default to `type=button`. We did this so
these close buttons did not submit the forms unintentionally. Without this change, you have to
manually add `type="button"` to these components if you wrap popup contents in a `form` element.
While this is a very unlikely that a `CloseButton` was intentionally used as an implicit submit
button, it is still a breaking change. This change effects the following components:

- `Popup.CloseIcon`
- `Dialog.CloseIcon`
- `Modal.CloseIcon`
- `Popup.CloseButton`
- `Dialog.CloseButton`
- `Modal.CloseButton`

Co-authored-by: @mannycarrera4 <mannycarrera4@users.noreply.github.com>