fix: Fail gracefully if localStorage is blocked in useModalityType (#1898)
There could be many reasons localStorage is blocked (content security policy, iframe cross origin, etc). `useModalityType` uses localStorage to avoid possible flashing of UI for users. But a blank screen due to an access denied error is worse, so we'll wrap `localStorage` access in a `try/catch` block to gracefully fallback to guessing when localStorage is not available.

[category:Components]