docs: Add example for a simple compound component (#1998)
Whenever I want to use the compound component pattern it is hard to find a nice easy example to copy. Users often need to find a "simple" existing component, or dive into the documentation. This PR adds a super simple example of how to extend an existing CKR component with a customized wrapper. The goal is to have a one click place to get a skeleton that can be customized.

[category:Documentation]