chore!: Hard deprecate Layout and Column components (#2018)
Fixes: #1994  

The `Layout` and `Column` components have been determined to be hard deprecated and fully removed from Canvas Kit as of v9. As they are not longer available to use, they can be replaced by more modern and efficient alternatives, such as `Grid`.

[category:Components]

### BREAKING CHANGES
We have hard deprecated the Layout and Column components and they have been fully removed from our codebase in v9. We recommend using the `Grid` component as an alternative.