feat!: Remove Stack, HStack and VStack (#2012)
This is a hard deprecate of `Stack`, `HStack`, `VStack` along with the props associated to it (`StackProps`, `HStackProps`, `VStackProps` and `StackStyleProps`).

Fixes: #1999 

We now have `gap` support since dropping support for IE and no longer need `Stack`, `HStack` and `VStack`.

[category:Components]

### BREAKING CHANGES
We've hard deprecated the `Stack`, `HStack` and `VStack` components. We recommend using the `Flex` component as a replacement and replacing the `spacing` prop with the `gap` prop. You can also run `npx @workday/canvas-kit-codemod softDeprecate/Stack` as well.