fix: Extract the correct ref with a styled component (#2122)
This change allows a styled component to be used as the default component in `createComponent`, `createContainer`, and `createSubcomponent`. Before the `ref` was not be extracted correctly. This mean the `ref` was incompatible with any other components.

[category:Components]

Release Note:
The following is now allowed:

```tsx
const StyledComponent = styled('button')({})

const MyComponnent = createComponent(StyledComponent)({
  Component(elemProps, ref, Element) {
    return <Box as={Element} ref={ref} {...elemProps} />
  }
})
```