feat(text-input): Add InputGroup component (#2182)
Add an `InputGroup` component to handle icons in a `TextInput` component.

There is already an `InputIconContainer` component that is not documented, but is limited to only right-aligned icons and doesn't handle RTL.

[category:Components]

Release Note:
`InputGroup` will replace `InputIconContainer`. `InputIconContainer` does not handle bidirectionality or icons at the start of an input. `InputIconContainer` will be deprecated and later removed in future versions.

Before:
```tsx
<InputIconContainer icon={<SystemIcon icon={exclamationCircleIcon} />} />
```

After
```tsx
<InputGroup>
  <InputGroup.Input />
  <InputGroup.InnerEnd>
    <SystemIcon icon={exclamationCircleIcon} />
  </InputGroup.InnerEnd>
</InputGroup>
```