chore(collection)!: Move initial cursor to roving focus (#2205)
Fixes: #1234 

Not all collections need an initial cursor position to function. This functionality is more suitable for roving tabindex and doesn't work as well for `aria-activedescendant`. This change moves the functionality out of the cursor model and into `useListItemRovingFocus`

[category:Components]

### BREAKING CHANGES
If you use `useListModel` to create AND need the initially registered item to be the cursor position AND do not use `useListItemRovingFocus`, you'll have to add logic to set the cursor to that first item in the list. If you use both `useListModel` and `useListItemRovingFocus`, there is not change in behavior.