fix: Focusable elements not always focused (#2204)
Some React component libraries (notably https://www.npmjs.com/package/react-select) have refs that are focusable but are not HTMLElements. Due to the ordering of the logic currently, these components are not focused correctly. 

Example: 
We have a Canvas Kit popup modal that sets initialFocusRef to a react-select component. Since it is typed differently than Canvas Kit expects here, it does not correctly receive keyboard focus and the keyboard focus remains on the background elements (behind the modal!). This is bad for accessibility.

[category:Components]

Co-authored-by: @NicholasBoll <nicholas.boll@gmail.com>