ci: Fix canary releases (#2228)
Lerna doesn't seem to support the `canary` option anymore when creating prereleases. The command doesn't work correctly for us anyways, so I'm removing it. From what I can gather (docs don't mention it and info is sparse), it is supposed to create a pre-id using previous tags and the number of commits between the last tag and the current `HEAD`. But this caused problems for us for a long time because we release with every commit on `master` which caused the commit count to repeat and caused the order of versions to be listed out of order. We started using CI build numbers to fix the problem [#1281](https://github.com/Workday/canvas-kit/pull/1281/commits/95b31c1ae82e9a0996996e03a41c86f68419c96e).

[category:Infrastructure]