ci: Support fast-forward-merges [skip-release] (#2232)
This change updates the forward merge action to support [fast-forward-only](https://git-scm.com/docs/git-merge#Documentation/git-merge.txt---ff-only). This is useful after we merge `prerelease/minor` when doing a minor release or after a major release. A fast-forward-only merge simply moves the `HEAD` of the branch since there is no divergence yet. This prevents all the merges in the history when it isn't necessary.

[category:Infrastructure]