fix: Add custom fallback placements modifier (#2200)
Fixes: #1100 
This fix is to ensure the Popper has the ability to dynamically calculate the available spaces around the trigger and display the Popper at the proper position provided by `fallbackPlacements` array.

With the new fix, Popper will dynamically check the available positions in the following order of **the initial placement**, **the opposite placement**, and **the rest of fallback placements** via the custom `fallbackPlacement` modifier.

[category:Components]

Release Note:
This fix adds additional fallback placements to prevent popups from rendering off screen. This behavior is different than before and might result in a popup rendering in an unexpected position. `fallbackPlacements` was added as a prop to `Popper` and `Popup.Popper` to allow you to override the default fallback placements (which are top/right/bottom/left in that order). This shouldn't be considered a breaking change unless your popup position needs to be restricted.

Co-authored-by: clare.huang <clare.huang@workday.com>
Co-authored-by: @NicholasBoll <nicholas.boll@gmail.com>