fix(collection): Default `T` to `any` to allow overriding children types (#2254)
Default `T` in the `ListProps` type signature to `any` to allow overriding of the `item` in the dynamic signature.

For example:

```tsx
<Menu items={items}>
  <Menu.List>
    {(item: Item) => <Menu.Item data-id={item.id}>{item.text}</Menu.Item>
  </Menu.List>
</Menu>
```

In this example, TypeScript will error on `item: Item` because `item` is defaulted to `unknown` which TypeScript complains is incompatible. We use `any` because there is no way to infer `item` in the render prop from the `items` from the `Menu` component. TypeScript doesn't do inference across JSX components.

[category:Components]