feat!: Remove Menu Preview (#2247)
Removal of the `DeprecatedMenu` and `DeprecatedMenuItem` component in Preview.

Fixes: #1993  

A compound component version is already in Main.

[category:Components]

Release Note:
This will not involve a codemod since the API has too many differences between the component being removed in Preview and the component in Main.

### BREAKING CHANGES
To move from the Menu component in Preview to the Menu compound component in Main, please read [this discussion](https://github.com/Workday/canvas-kit/discussions/2063).

Also, see [this](https://codesandbox.io/s/deprecatedmenu-and-menu-migration-gwi1ov) codesandbox that shows the differences between DeprecatedMenu and the Menu compound component.