fix(preview): Blur firing incorrectly onClick (#2304)
Fixes: #2303 

When you click to open the preview `<Select />` menu, onBlur event is immediately triggered. This is somewhat expected since the focus moves from the button to the menu, however is not desirable since many validation libraries use the blur event to check the field. This change blocks triggering the blur event unless the menu is closed.

[category:Components]