fix(popup): Include document.body in useReturnFocus bound check (#2351)
Fixes: #2349 

`useReturnFocus` has been updated to consider whether the element that received focus is visible on the screen. This includes checking if the element is visible in the scroll parent as well as in the viewport. Previously, focus would only be returned if the element was visible in its scroll parent but not necessarily in the viewport.

[category:Components]

Co-authored-by: @ebin-benny-wd <ebin.benny@workday.com>