feat: Add CSS styling function (#2273)
Implements: #2208 
Related: #1854

Discussion: #2265 

This change adds a styling function to use CSS-in-JS with a minimum developer runtime and a transform process that can turn CSS-in-JS into static styling.

[category:Components]

Release Note:
This change introduces an intermediate step towards static styling. The style function takes in statically analyzable style definitions with modifiers, variables, and static styles to create styling to be used in components. It can later be interpreted at build time to create CSS packages.