chore(button): Add CSS Variable fallbacks (#2372)
Add CSS Variable fallbacks so an environment is not required to add Canvas Tokens to the environment for the default theme. This makes it much easier for users to have sandbox environments like Storybook without Canvas Components looking broken. These fallbacks will need to be removed when the Style transform automatically adds fallbacks.

[category:Components]