fix(preview): Fix multiline asterisk on preview form field (#2436)
Preview Form field has the same multiline asterisk issue as the main line one, see #2217. This fixes the asterisk so it is always pinned to the end of the text.

[category:Components]

Release Note:
We removed a wrapping flex element so elemProps is now spread directly on the `<label>` element and the asterisk is at the end of the label.