fix: Default createStyles vars to empty object (#2488)
Allow booleans on modifiers in Stencils. We need to default the `vars` config in Stencils to allow the `MaybeBoolean` type to work. Otherwise we'll get `vars` defaulting to `Record<string, string> | Record<Record<string, string>` which won't be compatible if `modifiers` have a `boolean` type. This won't work: `<Record<string, string> & { someKey?: boolean }`. `Record<string, string>` is too overly permissive and we need to default `vars` to an empty object to avoid widening the type.

[category:Components]