fix: SidePanel Examples aria-labelledby Update (#2538)
Fixes: #2523  

For the Basic example, Alternate example, and Right-to-left example, they should display a hidden `<span>` with `labelProps` when in the collapsed state. This helps keep the `aria-labelledby` reference on the toggle button valid while collapsed.

[category:Documentation]

Release Note:
Updates the Basic example, Alternate example, and Right-to-left example to display a hidden <span/> element when in the collapsed state.

Co-authored-by: @josh-bagwell <44883293+josh-bagwell@users.noreply.github.com>