docs: Add interactive table examples to storybook (#2457)
Resolves: none

Integrating a checkbox labeled "select all" into a table's column header has been a "hot topic" in accessibility. This is primarily because screen readers will announce "select all" automatically each time users read any cell in column 1 of a table, that is used for selecting rows. This is admittedly confusing; the accessibility team researched an improvement to this experience.

[category:Documentation]

Co-authored-by: @alanbsmith <a.bax.smith@gmail.com>