fix: Fix keyframes transform for commonjs modules (#2572)
Fixes the style transform for keyframes in commonjs module export type. If the `keyframe` is a call expression node type, the `ts.factory.updateCallExpression` is used instead of creating a new one. Creating a new AST node has a negative effect on TypeScript's own `commonjs` module transformation where it looses access to the original parent node and doesn't transform the import correctly.

[category:Components]