fix: Forward numeric style prop transform handling to Emotion (#2588)
Fixes: #2587

The static style transform was always turning numbers into strings with a `px` suffix. This isn't correct and Emotion has extra logic to turn a number into the correct string. For example, an opacity doesn't have a unit. This fix updates the logic to forward numbers straight to Emotion to process which should give a correct result.

[category:Infrastructure]