fix(button): Fix icon colors in buttons and SegmentedControl (#2623)
Fixes: #2622 

This continues with the intended change to icons in `SegmentedControl` that were broken with icons with background layers. The intent was to treat all icon fill layers the same in `SegmentedControl` and we failed to do that properly. This change simplifies the icon fill selectors and ensures the icons always look as intended and don't require guess work from icon designers.

[category:Components]

Release Note:
This change may trigger a visual regression with certain icons in buttons and `SegmentedControl` buttons. This change is intentional and fixes an accidental regression between v9 and v10. There should be no other breaking changes.