fix(styling): Fix variables in stencil config (#2624)
Fixes variables used inside a stencil. Before this used to cause problems when `color` is expected to be a string:
```tsx
const myStencil = createStencil({
  vars: {
    color
  },
  base({color}) {
    color // `string | undefined`, but should aways be a `string`
  }
})
```

This change also introduces the concept of uninitialized stencil variables. It always worked, but I've added documentation and support in the static style parser.

I also fixed a but where JSDoc of a stencil variable did not get forwarded properly.

[category:Styling]