fix(stencil): TypeScript support for variable/modifier with same key (#2635)
Fix TypeScript types when the variable and modifiers share the same keys. This is supported without issue at runtime, but TypeScript would throw an error. Now if the keys are the same, the types will include the modifier keys, but allow any string.

For example, the following is now supported:
```tsx
const myStencil = createStencil({
  vars: {
    width: '10px',
  },
  base({width}): {
    return {
      width: width
    }
  },
  modifiers: {
    width: {
      zero: {
        width: '0', // overrides base styles
      },
    },
  },
});

myStencil({width: 'zero'}); // `'zero'` is part of autocomplete
myStencil({width: '10px'}); // width also accepts a string
```

[category:Styling]