feat: Add flex styles to mergeStyles (#2731)
Flex style props are not part of `mergeStyles`. This will cause breaking changes if we remove style props from a component when an application uses flex style props. For example,

```tsx
<ListBox flexDirection="row">...</ListBox>
```

If we used a `flexDirection` style prop in v10, but move it to a Stencil in v11, the `flexDirection="row"` will no longer work. This is because `mergeStyles` is covering all style props except for flex style props like `flexDirection`. The `flexDirection` prop will be prop drilled all the way down to the `Flex` layout component which will now be applied under the `flexDirection: 'column'` of the Stencil, changing the behavior of the style prop. It is not our intention to break style props in v11, so the `mergeStyles` should handle all style props so that passing style props this way always overrides.

[category:Components]