chore: Improve CSS Kit selectors (#2751)
Improve CSS Kit selectors by dropping the prefix for modifiers.

[category:Styling

--

**HTML**:
```html
// before 
<button class="cnvs-button cnvs-button-tertiary cnvs-button--size-large cnvs-button-tertiary--size-large cnvs-button--icon-position-only cnvs-button-tertiary--icon-position-only">

// after
<button class="cnvs-button cnvs-button-tertiary size-large icon-position-only">
```

**CSS**:
```css
// before 
.cnvs-button--size-large.cnvs-button--icon-position-only {
  // base button compound modifier styles
}

.cnvs-button-tertiary.cnvs-button--size-large.cnvs-button--icon-position-only {
  // tertiary button compound modifier styles
}

// after
.cnvs-button.size-large.icon-position-only {
  // base button compound modifier styles
}

.cnvs-button-tertiary.size-large.icon-position-only {
  // tertiary button compound modifier styles
}
```

It wasn't clear where prefixes were coming from (`BaseButton` or `TertiaryButton`). The transform tries to track which modifiers are from the base and which are from extension, but adding new modifiers take on the extended name. For a CSS Kit HTML author, they'll have to keep track which modifiers come from which extension level. It was easy to make a mistake.

This change ensures the modifiers are semantic with no prefix so developers don't have to care where the modifiers come from, just the name of them.]