fix: Prevent undefined from overriding model defaultConfig (#2766)
Fixes: #2756

Models use the object spread operator to allow `config` to override `defaultConfig`. If `undefined` is passed to a key, it would cause `undefined` to be assigned to config even if the model didn't allow for that. Instead of using an object spread, `config` keys are looped over to check for `undefined` values.

[category:Components]