fix(combobox): Use correct state for aria-selected (#2849)
Fixes 2860.  `aria-selected` for combobox menu items. The cursor was used to determine `aria-selected` instead of the selected item. This changes the visuals of the selected vs focus state, but that issue is covered in #2828.

[category:Components]

Release Note:
This change fixes `aria-selected` in `Combobox.Menu.Item` components, but this does change the visuals of what is considered "selected". If you have any visual tests that have a screenshot of a selected state, the visual regression will have to be updated. The same is true for DOM-based snapshot tests. `aria-selected="true"` will now be added when an item is selected and not just when the virtual cursor is on the item. If your snapshot captures this DOM state, the snapshot will have to be updated.