refactor(select): Fix all syncing issues with the Select component (#2983)
The `useComboboxInputConstrained` hook was created for Comboboxes that can only have a value from within a set of defined options. The value is constrained by the options passed to it. `useComboboxInputConstrained` assumes there will be multiple inputs where one is presented to the user and the other is for the form/server. The hook keeps the model in sync with external ways to manipulate the input: React `value` prop ([controlled inputs](https://react.dev/reference/react-dom/components/input#controlling-an-input-with-a-state-variable)), browser/extension autofill, and testing tools.

`Select` was updated to use `useComboboxInputConstrained` which fixes several bugs at once:
- Fixes https://github.com/Workday/canvas-kit/issues/2888
- Fixes https://github.com/Workday/canvas-kit/issues/2675
- Fixes https://github.com/Workday/canvas-kit/issues/2616
- Fixes https://github.com/Workday/canvas-kit/issues/2533

[category:Components]