feat: Add backgroundColor var on LoadingDots (#2991)
Fixes: #2990 

https://canvas.workdaydesign.com/components/indicators/loading-dots/ shows a circle variant for use on gray/dark/image-based backgrounds. Our LoadingDots component styling needed some TLC to allow consumers to set the background color to match this pattern, and our Storybook entries can be updated to show how to achieve this variant.

I also noticed some links to docs were broken in the codebase so I updated that too.

[category:Components
Documentation]

Release Note:
Adds capability to LoadingDots to accept a `loadingDotColor` prop to change loading dots color, with the capability to override that var with a stencil - intended for use with the circle variant for dark backgrounds/images, which is now shown in Storybook with Custom Shape story.
Also adds a `animationDurationMs` prop in the same pattern, with a Custom Color and Animation story in storybook demonstrating the use of both these new props.

Co-authored-by: @mannycarrera4 <mannycarrera4@users.noreply.github.com>
Co-authored-by: manuel.carrera <manuel.carrera@workday.com>
Co-authored-by: @NicholasBoll <nicholas.boll@gmail.com>
Co-authored-by: @alanbsmith <alanbsmith@users.noreply.github.com>
Co-authored-by: @alanbsmith <a.bax.smith@gmail.com>
Co-authored-by: @mannycarrera4 <mannycarrera4@users.noreply.github.com>