feat(collection): Add removable support (#3036)
Fixes: #3025

Adds `remove` event to the `ListModel`. An `onRemove` config should be added to dynamic lists to remove the item from the collection.

The `MultiSelect.Input` uses this new remove event to handle removing items from the Selected pill list when the user uses the “Delete” key. Focus is managed by the collection system when an item is removed.

[category:Components]