fix: Resolve recast dep from jscodeshift to remove extra parens (#3083)
Fixes: https://github.com/Workday/canvas-kit/issues/3084

Based on an [open issue in jscodeshift](https://github.com/facebook/jscodeshift/issues/534), `recast` seems to be adding an extra parens in the return call in our unit tests. One of the suggested solutions was to downgrade `recast`. In v12, this has been fixed, but adding this resolution for v11.

[category:Dependencies]

Co-authored-by: manuel.carrera <manuel.carrera@workday.com>