feat: Add new `description` variant to Tooltip and deprecate `describe` variant for better accessibility. (#2814)
Fixes: #2797

[category:Components]

Release Note:
- We've added a new `description` variant and **deprecated** the `describe` variant.

The `describe` variant of the tool tip relies on `aria-describedby` attribute to assign the tool tip text to the accessible description of the target element. This only works when the tool tip is visible and present in the DOM, which only occurs when the element has focus. Screen readers can access web page elements without necessarily using keyboard focus, therefore, this variant didn't appear to always work reliably for screen reader users. 

The preferred type to use is `description` to provide better accessible support.

Co-authored-by: manuel.carrera <manuel.carrera@workday.com>
Co-authored-by: @mannycarrera4 <mannycarrera4@users.noreply.github.com>