feat: Add New Component `BaseTable` (#3126)
Fixes: #3048  

Add a new base component for the Table component: `BaseTable`. This will be a standard Html `Table` without any  `CSS Grid` features. This will be the "base" level component that will be used to build other Tables.

[category:Components]

Release Note:
Addition of a new base level component: `BaseTable`. This will be a standard HTML `Table` with no `CSS Grid` features.